+++ lib/msf/core/exploit.rb
@@ -1,7 +1,6 @@
 # -*- coding: binary -*-
 require 'msf/core'
 require 'msf/core/module'
-require 'pry'
 
 module Msf
 
@@ -757,15 +756,7 @@ class Exploit < Msf::Module
     # what not?
     return false if !compatible?(pi)
 
-    if defined?(@needs_cleanup) && @needs_cleanup
-      puts("\n#{self.name} needs cleanup!")
-    else
-      puts("#{self.name} does not need cleanup!")
-    end
-    puts("#{pi.name} cleans_up: #{pi.can_cleanup}")
     if defined?(@needs_cleanup) && @needs_cleanup && !pi.can_cleanup
-      puts("module #{self.name} needs cleanup!")
-      puts("payload #{pi.name} cannot cleanup!")
       return false
     end
