+++ lib/msf/core/exploit/file_dropper.rb
@@ -1,17 +1,15 @@
 # -*- coding: binary -*-
-require 'pry'
 
 module Msf
 module Exploit::FileDropper
-
 
   def initialize(info = {})
     super
 
     @needs_cleanup = true
-    puts("File Dropper!")
     @dropped_files = []
     @dropped_dirs = []
+
     register_advanced_options(
       [
         OptInt.new('FileDropperDelay', [false, 'Delay in seconds before attempting cleanup'])
