+++ lib/msf/core/modules/metadata/obj.rb
@@ -53,8 +53,6 @@ class Obj
   attr_reader :default_credential
   # @return [Hash]
   attr_reader :notes
-  # @return [Boolean]
-  attr_reader :cleanup_required
 
   def initialize(module_instance, obj_hash = nil)
     unless obj_hash.nil?
@@ -115,7 +113,6 @@ class Obj
   # Returns the JSON representation of the module metadata
   #
   def to_json(*args)
-    puts("to_json")
     {
       'name'               => @name,
       'fullname'           => @fullname,
