+++ lib/msf/core/payload.rb
@@ -659,6 +659,7 @@ protected
 
     return raw
   end
+
   ##
   #
   # Custom merge operations for payloads
