+++ lib/msf/core/payload/pingback.rb
@@ -3,7 +3,6 @@
 require 'msf/core'
 require 'msf/core/module/platform'
 require 'rex/text'
-require 'pry'
 
 #
 # This class provides methods for calculating, extracting, and parsing
@@ -11,22 +10,40 @@ require 'pry'
 #
 module Msf::Payload::Pingback
 
-
-  #
-  # Constants
-  #
-
   #
   # Instance methods
   #
+  #
+  # Generates a URI with a given checksum and optionally with an embedded UUID if
+  # the desired length can accommodate it.
+  #
+  # @param mode [Symbol] The type of checksum to generate (:connect, :init_native, :init_python, :init_java)
+  # @param len [Integer] The length of the URI not including the leading slash, optionally nil for random
+  # @return [String] A URI with a leading slash that hashes to the checksum, with an optional UUID
+  #
+
+  # Generate a Payload UUID
+  def generate_pingback_uuid
+    self.pingback_uuid ||= SecureRandom.uuid()
+    print_status("PingbackUUID = #{self.pingback_uuid}")
+    if framework.db.active
+      print_status("Writing UUID #{datastore['PingbackUUID']} to database...")
+      Mdm::Payload.create!(name: datastore['PayloadUUIDName'],
+                           uuid: datastore['PingbackUUID'].gsub('-',''),
+                           description: 'pingback',
+                           platform: platform.platforms.first.realname.downcase)
+    else
+      print_warning("Unable to save UUID to database -- database support not active")
+    end
+    self.pingback_uuid
+  end
 
   def initialize(info = {})
     ret = super(info)
-    puts("In pingback costructor")
     self.can_cleanup = false
-    self.uuid ||= SecureRandom.uuid()
+    self
   end
 
-  attr_accessor :uuid
+  attr_accessor :pingback_uuid
   attr_accessor :can_cleanup
 end
